accountapi/requirements.txt
django~=5.0.2
djangorestframework~=3.14.0
mysqlclient~=2.2.4
django-rest-swagger~=2.2.0
drf-yasg~=1.21.7
setuptools~=69.1.0
py-eureka-client

docker~=7.0.0
py_eureka_client~=0.11.9